+++ pkg/analysis_server/test/abstract_context.dart
@@ -90,7 +90,7 @@ class AbstractContextTest with ResourceProviderMixin {
     var analysisContext = contextFor(testPackageRootPath);
     var files = analysisContext.contextRoot.analyzedFiles().toList();
     for (var path in files) {
-      await analysisContext.currentSession.getResolvedUnit(path);
+      await analysisContext.currentSession.getResolvedUnit2(path);
     }
   }
 
@@ -171,7 +171,8 @@ class AbstractContextTest with ResourceProviderMixin {
 
   Future<ResolvedUnitResult> resolveFile(String path) async {
     path = convertPath(path);
-    return contextFor(path).currentSession.getResolvedUnit(path);
+    var session = contextFor(path).currentSession;
+    return await session.getResolvedUnit2(path) as ResolvedUnitResult;
   }
 
   @mustCallSuper