+++ pkg/analysis_server/test/services/correction/organize_directives_test.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:analysis_server/src/services/correction/organize_imports.dart';
+import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/error/error.dart';
 import 'package:analyzer_plugin/protocol/protocol_common.dart'
     hide AnalysisError;
@@ -581,7 +582,7 @@ import 'package:b/a.dart';''');
 
   Future<void> _computeUnitAndErrors(String code) async {
     addTestSource(code);
-    var result = await session.getResolvedUnit(testFile);
+    var result = await session.getResolvedUnit2(testFile) as ResolvedUnitResult;
     testUnit = result.unit!;
     testErrors = result.errors;
   }