+++ pkg/analysis_server/test/services/refactoring/move_file_test.dart
@@ -4,6 +4,7 @@
 
 import 'package:analysis_server/src/protocol_server.dart';
 import 'package:analysis_server/src/services/refactoring/refactoring.dart';
+import 'package:analyzer/dart/analysis/results.dart';
 import 'package:test/test.dart';
 import 'package:test_reflective_loader/test_reflective_loader.dart';
 
@@ -65,7 +66,8 @@ import 'package:test/old_name.dart';
     // Since the file being refactored isn't the test source, we set the
     // testAnalysisResult manually here, the path is referenced through the
     // referenced File object to run on Windows:
-    testAnalysisResult = await session.getResolvedUnit(file.path);
+    testAnalysisResult =
+        await session.getResolvedUnit2(file.path) as ResolvedUnitResult;
 
     _createRefactoring('/home/test/lib/222/new_name.dart', oldFile: file.path);
     await _assertSuccessfulRefactoring();
@@ -87,7 +89,8 @@ import 'package:test0.test1.test2/111/name.dart';
     // Since the file being refactored isn't the test source, we set the
     // testAnalysisResult manually here, the path is referenced through the
     // referenced File object to run on Windows:
-    testAnalysisResult = await session.getResolvedUnit(file.path);
+    testAnalysisResult =
+        await session.getResolvedUnit2(file.path) as ResolvedUnitResult;
 
     _createRefactoring('/home/test0/test1/test3/lib/111/name.dart',
         oldFile: file.path);
@@ -110,7 +113,8 @@ import 'package:test0.test1.test2/111/name.dart';
     // Since the file being refactored isn't the test source, we set the
     // testAnalysisResult manually here, the path is referenced through the
     // referenced File object to run on Windows:
-    testAnalysisResult = await session.getResolvedUnit(file.path);
+    testAnalysisResult =
+        await session.getResolvedUnit2(file.path) as ResolvedUnitResult;
 
     _createRefactoring('/home/test0/test1/test2/test3/lib/111/name.dart',
         oldFile: file.path);
@@ -133,7 +137,8 @@ import 'package:test0.test1.test2/111/name.dart';
     // Since the file being refactored isn't the test source, we set the
     // testAnalysisResult manually here, the path is referenced through the
     // referenced File object to run on Windows:
-    testAnalysisResult = await session.getResolvedUnit(file.path);
+    testAnalysisResult =
+        await session.getResolvedUnit2(file.path) as ResolvedUnitResult;
 
     _createRefactoring('/home/test0/test1/lib/111/name.dart',
         oldFile: file.path);
@@ -154,7 +159,8 @@ import 'package:test/111/old_name.dart';
     // Since the file being refactored isn't the test source, we set the
     // testAnalysisResult manually here, the path is referenced through the
     // referenced File object to run on Windows:
-    testAnalysisResult = await session.getResolvedUnit(file.path);
+    testAnalysisResult =
+        await session.getResolvedUnit2(file.path) as ResolvedUnitResult;
 
     _createRefactoring('/home/test/lib/222/new_name.dart', oldFile: file.path);
     await _assertSuccessfulRefactoring();
@@ -174,7 +180,8 @@ import 'package:test/222/old_name.dart';
     // Since the file being refactored isn't the test source, we set the
     // testAnalysisResult manually here, the path is referenced through the
     // referenced File object to run on Windows:
-    testAnalysisResult = await session.getResolvedUnit(file.path);
+    testAnalysisResult =
+        await session.getResolvedUnit2(file.path) as ResolvedUnitResult;
 
     _createRefactoring('/home/test/lib/new_name.dart', oldFile: file.path);
     await _assertSuccessfulRefactoring();