+++ pkg/nnbd_migration/test/front_end/info_builder_test.dart
@@ -25,7 +25,7 @@ void main() {
 @reflectiveTest
 class BuildEnclosingMemberDescriptionTest extends AbstractAnalysisTest {
   Future<ResolvedUnitResult> resolveTestFile() async {
-    return await session.getResolvedUnit(testFile);
+    return await session.getResolvedUnit2(testFile) as ResolvedUnitResult;
   }
 
   Future<void> test_classConstructor_named() async {